+++ lib/components/layouts/Sidebar/constant.js
@@ -1,0 +1,4 @@
+export const TOOLTIP_STYLES = {
+  default: "default",
+  featured: "featured",
+};
